+++ src/pages/main/googlemaps/index.tsx
@@ -1,17 +1,16 @@
 import { useEffect, useState } from 'react';
-import { GoogleMap, InfoWindow, Marker } from '@react-google-maps/api';
+import { GoogleMap, Marker } from '@react-google-maps/api';
 import SearchBox from '@/components/GoogleMap/searchBox';
 import GMap from '@/components/GoogleMap/gmap';
 import caxios from '@/util/caxios';
-import CountryInfo from '@/components/GoogleMap/countryInfo';
 import CountryInfoArea from './countryInfoArea';
+import CovidCountryInfo from '@/components/GoogleMap/covidCountryInfo';
 
 const defaultCenter: { lat: number; lng: number } = {
   lat: 37,
   lng: 127,
 };
 const GoogleMaps = () => {
-  const [selected, setSelected] = useState<any>(null); // marker를 선택하면 그 위치에 대한 정보를 담는다.
   const [currentPosition, setCurrentPosition] = useState<any>({}); // marker를 드래그해서 마지막에 놓은 위치의 정보를 담는다.
   const [locMarker, setLocMarker] = useState<MapAPI.LockMarkerWithLogin>(); // 검색한 위치의 marker를 표시한다.
   const [locations, setLocations] = useState<any>([]);
@@ -75,33 +74,12 @@ const GoogleMaps = () => {
     });
   };
 
-  // Marker를 클릭하면 정보를 가지고온다.
-  const onSelect = (val: any) => {
-    console.log('onselect', val);
-    setSelected(val);
-  };
-
   // Marker를 드래그하고 내려놓은 곳의 정보를 가져온다.
   const onMarkerDragEnd = (e: any) => {
     console.log(e);
     const lat = e.latLng.lat();
     const lng = e.latLng.lng();
     setCurrentPosition({ lat, lng });
-  };
-
-  // Marker icon변경
-  const setCovidIcon = (loc: any): google.maps.Icon => {
-    let alarmLvl: number = 0;
-
-    if (loc?.alarmLvl === 1) alarmLvl = 1;
-    else if (loc?.alarmLvl === 2) alarmLvl = 2;
-    else if (loc?.alarmLvl === 3) alarmLvl = 3;
-    else alarmLvl = 4;
-
-    return {
-      url: require(`../../../components/GoogleMap/asset/${alarmLvl}.png`),
-      scaledSize: new google.maps.Size(20, 20),
-    };
   };
 
   const parseFloat = (val: string) => {
@@ -111,71 +89,36 @@ const GoogleMaps = () => {
   const setMarkerTypeIcon = (val: any): google.maps.Icon => {
     let iconType: string = '';
 
-    if (val?.isTravelInfo) iconType = 'airplain';
-    else iconType = 'pin';
+    if (val?.isTravelInfo) iconType = 'airplain-1';
+    else iconType = 'pin2-1';
 
     return {
       url: require(`../../../components/GoogleMap/asset/${iconType}.png`),
-      scaledSize: new google.maps.Size(40, 50),
+      // scaledSize: new window.google.maps.Size(40, 40),
     };
   };
+
   return (
     <div>
       <GMap>
         <GoogleMap
           mapContainerStyle={{ height: '85vh', width: '100%' }}
-          zoom={locMarker !== undefined ? locMarker[0].zoom : 4}
-          center={locMarker !== undefined ? locMarker[0].location : defaultCenter}
+          zoom={locMarker !== undefined ? locMarker[locMarker.length - 1].zoom : 4}
+          center={
+            locMarker !== undefined ? locMarker[locMarker.length - 1].location : defaultCenter
+          }
         >
-          <SearchBox setLocMarker={setLocMarker} />
-          {locations?.map((val: any, index: number) => {
-            return (
-              <Marker
-                key={val?.countryIsoAlp2}
-                position={{ lat: val?.lat, lng: val?.lng }}
-                clickable={true}
-                onClick={() => onSelect(val)}
-                visible={true}
-                icon={setCovidIcon(val)}
-              />
-            );
-          })}
-          {/* 검색한 위치가 있을 시 표시한다. */}
-          {/* {locMarker && locMarker.isSearched && <MarkerF position={locMarker.location} />} */}
+          <SearchBox locMarker={locMarker} setLocMarker={setLocMarker} />
 
-          {/* 저장된 여행지의 정보 */}
-          {locMarker !== undefined &&
-            locMarker.map((val: any, index: number) => {
-              return (
-                <Marker
-                  key={index}
-                  position={val.location}
-                  icon={{
-                    url: require(`../../../components/GoogleMap/asset/pin2.png`),
-                    size: new google.maps.Size(40, 50),
-                    // scaledSize: new google.maps.Size(40, 50),
-                  }}
-                />
-              );
-            })}
+          {/* 지도에 코로나 단계별로 마커를 표시해준다. */}
+          <CovidCountryInfo locations={locations} />
+
+          {/* 검색후 나타난 결과를 지도에 표시 및 저장된 여행지의 정보 */}
+          {locMarker?.map((val: any, index: number) => {
+            return <Marker key={index} position={val.location} icon={setMarkerTypeIcon(val)} />;
+          })}
 
           <CountryInfoArea loc={locations} />
-          {
-            // 툴팁
-            selected && (
-              <InfoWindow
-                position={{ lat: selected?.lat, lng: selected?.lng }}
-                onCloseClick={() => onSelect(null)}
-              >
-                <div>
-                  <p>국가: {selected?.countryNm}</p>
-                  <p>
-                    국가 정보: {selected?.txtOriginCn != null ? selected?.txtOriginCn : '정보없음'}
-                  </p>
-                </div>
-              </InfoWindow>
-            )
-          }
         </GoogleMap>
       </GMap>
     </div>
