+++ src/pages/nologin/googlemaps/index.tsx
@@ -1,13 +1,12 @@
 import { useEffect, useState } from 'react';
-import { GoogleMap, InfoWindow, InfoWindowF, Marker } from '@react-google-maps/api';
+import { GoogleMap, Marker } from '@react-google-maps/api';
 import SearchBox from '@/components/GoogleMap/searchBox';
 import GMap from '@/components/GoogleMap/gmap';
 import caxios from '@/util/caxios';
-import CountryInfo from '@/components/GoogleMap/countryInfo';
 import CountryInfoArea from './countryInfoArea';
+import CovidCountryInfo from '@/components/GoogleMap/covidCountryInfo';
 
 const GoogleMaps = () => {
-  const [selected, setSelected] = useState<any>(null); // marker를 선택하면 그 위치에 대한 정보를 담는다.
   const [currentPosition, setCurrentPosition] = useState<any>({}); // marker를 드래그해서 마지막에 놓은 위치의 정보를 담는다.
   const [locMarker, setLocMarker] = useState<MapAPI.LockMarker>(); // 검색한 위치의 marker를 표시한다.
   const [locations, setLocations] = useState<any>([]);
@@ -52,33 +51,12 @@ const GoogleMaps = () => {
     });
   };
 
-  // Marker를 클릭하면 정보를 가지고온다.
-  const onSelect = (val: any) => {
-    console.log('onselect', val);
-    setSelected(val);
-  };
-
   // Marker를 드래그하고 내려놓은 곳의 정보를 가져온다.
   const onMarkerDragEnd = (e: any) => {
     console.log(e);
     const lat = e.latLng.lat();
     const lng = e.latLng.lng();
     setCurrentPosition({ lat, lng });
-  };
-
-  // Marker icon변경
-  const setIcon = (loc: any): google.maps.Icon => {
-    let alarmLvl: number = 0;
-
-    if (loc?.alarmLvl === 1) alarmLvl = 1;
-    else if (loc?.alarmLvl === 2) alarmLvl = 2;
-    else if (loc?.alarmLvl === 3) alarmLvl = 3;
-    else alarmLvl = 4;
-
-    return {
-      url: require(`../../../components/GoogleMap/asset/${alarmLvl}.png`),
-      scaledSize: new google.maps.Size(20, 20),
-    };
   };
 
   const parseFloat = (val: string) => {
@@ -93,42 +71,16 @@ const GoogleMaps = () => {
           zoom={locMarker?.zoom}
           center={locMarker?.location}
         >
+          {/* 검색창 */}
           <SearchBox setLocMarker={setLocMarker} />
 
-          {locations?.map((val: any, index: number) => {
-            return (
-              <Marker
-                key={val?.countryIsoAlp2}
-                position={{ lat: val?.lat, lng: val?.lng }}
-                clickable={true}
-                onClick={() => onSelect(val)}
-                visible={true}
-                icon={setIcon(val)}
-              />
-            );
-          })}
+          {/* 지도에 코로나 단계별로 마커를 표시해준다. */}
+          <CovidCountryInfo locations={locations} />
 
           {/* 검색한 위치가 있을 시 표시한다. */}
           {locMarker && locMarker.isSearched && <Marker position={locMarker.location} />}
 
           <CountryInfoArea loc={locations} />
-
-          {
-            // 툴팁
-            selected && (
-              <InfoWindow
-                position={{ lat: selected?.lat, lng: selected?.lng }}
-                onCloseClick={() => onSelect(null)}
-              >
-                <div>
-                  <p>국가: {selected?.countryNm}</p>
-                  <p>
-                    국가 정보: {selected?.txtOriginCn != null ? selected?.txtOriginCn : '정보없음'}
-                  </p>
-                </div>
-              </InfoWindow>
-            )
-          }
         </GoogleMap>
       </GMap>
     </div>
